Significant refactor to support pluggable effects.
