+++ CMakeLists.txt
@@ -12,6 +12,7 @@ set(CMAKE_CXX_STANDARD 17)
 pico_sdk_init()
 
 include(bluetooth/bluetooth.cmake)
+include(effect/rainbow_fft.cmake)
 
 if(DISPLAY_PATH AND EXISTS ${CMAKE_CURRENT_LIST_DIR}/${DISPLAY_PATH})
 include(${CMAKE_CURRENT_LIST_DIR}/${DISPLAY_PATH})
@@ -24,17 +25,18 @@ add_executable(${NAME}
     ${CMAKE_CURRENT_LIST_DIR}/src/main.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/a2dp_sink.cpp
     ${CMAKE_CURRENT_LIST_DIR}/src/btstack_audio_pico.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/src/fixed_fft.cpp
 )
 
 target_include_directories(${NAME} PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}
+    ${CMAKE_CURRENT_LIST_DIR}/effect
 )
 
 target_link_libraries(${NAME}
     picow_bt_example_common
     pico_audio_i2s
     display
+    rainbow_fft
 )
 
 message(WARNING "Display: ${DISPLAY_NAME}")
